refactor: create the components page structure
