+++ src/components/pages/Form.tsx
@@ -1,9 +1,9 @@
-import Form from "@/components/form";
+import ActivitiesForm from "@/components/ActivitiesForm";
 
 export default function FormPage({ user_id }: { user_id: string }) {
   return (
     <div>
-      <Form user_id={user_id} />
+      <ActivitiesForm user_id={user_id} />
     </div>
   );
 }
